+++ application/modules/cards.py
@@ -1,57 +1,6 @@
 # -*- coding: utf-8 -*-
 from flask import render_template, redirect, url_for, flash
-from config import get_db
-
-def tuple_to_list(tuple_data):
-  return [list(tuple_row) for tuple_row in tuple_data]
-
-# def get_cards():
-#   cursor = get_db().cursor()
-#   SQL = """
-#   SELECT
-#   employee.first_name, employee.last_name, card.personnel_number
-#   FROM employee
-#   INNER JOIN card
-#   WHERE employee_id = employee.id;
-#   """
-#   cursor.execute(SQL)
-#   data = cursor.fetchall()
-#   cursor.close()
-#   data = tuple_to_list(data)
-#   for i, row in enumerate(data):
-#     data[i][0] = row[0].decode("utf-8")
-#   return render_template("cards.html", title = u"Личные карточки учета кадров", cards = data)
-
-# def get_card(card_id):
-#   card = None
-#   return card
-
-# def create_card(data):
-#   return True
-
-# def update_card(card_id, data):
-#   return True
-
-# def delete_card(card_id):
-#   return True
-
-def get_data_from_db(SQL_QUERY):
-  cursor = get_db().cursor()
-  cursor.execute(SQL_QUERY)
-  data = cursor.fetchall()
-  cursor.close()
-  return data
-
-
-def set_data_to_db(SQL_QUERY):
-  db = get_db()
-  cursor = db.cursor()
-  cursor.execute(SQL_QUERY)
-  db.commit()
-  data = cursor.fetchone()
-  cursor.close()
-  return data
-
+from utilites import *
 
 employee_contextmenu = {
   "module": {
@@ -79,12 +28,19 @@ employee_decodes = {
 def employees():
   QUERY = "SELECT id, first_name, last_name FROM employee;"
   data = get_data_from_db(QUERY)
+  employee_actions = [
+    dict(
+      name = "add",
+      caption = "Добавить",
+    ),
+  ]
   return render_template("table.html",
                   title = u"Сотрудники предприятия (базовая таблица)",
                   data = data, 
                   header = [u"#", u"Имя", u"Фамилия"],
                   contextmenu = employee_contextmenu,
                   decode = employee_decodes,
+                  actions = employee_actions,
                   )
 
 
@@ -138,15 +94,35 @@ def employee_form(_id, data = None):
       return redirect(url_for('index'))
     first_name = data["first_name"]
     last_name = data["last_name"]
-    data = set_data_to_db(UPDATE_QUERY%(first_name, last_name, _id))
-    # try:
-    #   data = set_data_to_db(UPDATE_QUERY%(first_name, last_name, _id))
-    # except:
-    #   flash(u"Ошибка в процессе изменения данных")
-    # else:
-    #   flash(u"Данные успешно изменены")
+    try Exception as err:
+      data = set_data_to_db(UPDATE_QUERY%(first_name, last_name, _id))
+    except:
+      flash(u"Ошибка в процессе изменения данных")
+      flash(str(err))
+    else:
+      flash(u"Данные успешно изменены")
     return redirect(url_for('index'))
 
+
+def add_employee(data = None):
+  INSERT_QUERY = "INSERT INTO "
+  if data is None:
+    return render_template("form.html", 
+                              items = employee,
+                              title = u"Основная информация о сотруднике",
+                              decode = employee_decodes,
+                            )
+  else:
+    first_name = data["first_name"]
+    last_name = data["last_name"]
+    try Exception as err:
+      data = set_data_to_db(UPDATE_QUERY%(first_name, last_name, _id))
+    except:
+      flash(u"Ошибка в процессе изменения данных")
+      flash(str(err))
+    else:
+      flash(u"Данные успешно изменены")
+    return redirect(url_for('index'))
 
 """
 CREATE TABLE `employee` (
