+++ application/modules/utilites.py
@@ -1,0 +1,24 @@
+# -*- coding: utf-8 -*-
+from config import get_db
+
+
+def tuple_to_list(tuple_data):
+  return [list(tuple_row) for tuple_row in tuple_data]
+
+
+def get_data_from_db(SQL_QUERY):
+  cursor = get_db().cursor()
+  cursor.execute(SQL_QUERY)
+  data = cursor.fetchall()
+  cursor.close()
+  return data
+
+
+def set_data_to_db(SQL_QUERY):
+  db = get_db()
+  cursor = db.cursor()
+  cursor.execute(SQL_QUERY)
+  db.commit()
+  data = cursor.fetchone()
+  cursor.close()
+  return data
